requirements.txt
pandas
streamlit
numpy
Pillow
pytesseract
opencv-python-headless